+++ iridauploader/parsers/parsers.py
@@ -42,9 +42,18 @@ def parser_factory(parser_type):
     if parser_type == "nextseq":
         logging.debug("Creating nextseq parser")
         return nextseq.Parser(parser_type_name=parser_type)
-    if parser_type == "nextseq_nml":
+    if parser_type == "nextseq_nml_classic":
         logging.debug("Creating nml custom nextseq parser")
-        return nextseq.Parser(parser_type_name=parser_type, sample_sheet_override='SampleSheetClassic.csv')
+        return nextseq.Parser(
+            parser_type_name=parser_type,
+            sample_sheet_override='SampleSheetClassic.csv'
+        )
+    if parser_type == "nextseq_nml_dual_upload":
+        logging.debug("Creating nml custom nextseq parser with extra required files")
+        return nextseq.Parser(
+            parser_type_name=parser_type,
+            additional_required_files=['RunMetadata.csv']
+        )
     if parser_type == "nextseq2k_nml":
         logging.debug("Creating nextseq2k_nml parser")
         return nextseq2k_nml.Parser(parser_type_name=parser_type)